Updates all get_base_turf usages to get_base_turf_by_area. This means that instead of the base turf being defined by the z-level, it's now defined by area and defaulted by z-level.
Updates all "shuttle" and "solar" areas to have a base_turf of space.
Updates docking arms 1, 2, and 3 on the main asteroid level to have a base_turf of space.

Other z-levels may still need updating.

This change means that all actions which destroy or dismantle a tile - for example bombs, a singularity, a shuttle leaving, or deconstruction - now leave these new areas as space not asteroid. Future mapping efforts may need to take this into account.