Unlimited Welder Fix
Fixes a bug where welders are infinite use.  With this, the Welder will not turn on unless it has at least 1u of welder fuel.  Previously, it worked on 'not zero', which allowed it to weld endlessly with fractions of 1u inside (as it was not Zero, but could not deduct 1 from it when used).