Upgrades default suit cooler cell
Considering suit coolers are dedicated EVA equipment instead of emergency equipment, it's strange that they come equipped with a cell that lasts a handful of minutes in operation instead of a duration more comparable to an oxygen tank's.

This upgrades their default cell so that FBPs can grab, combine, and go just like other crewmembers.