Adds accuracy penalties for certain detrimental states. Being blind gives an additional 75% chance to miss. Having blurry eyes gives an additional 15% chance to miss. Being confused gives an additional 30% chance to miss.
This was done so that if the PR for flashes to apply these effects instead of a stun are merged, they would hopefully still be useful if someone is trying to robust you.  Melee and ranged penalties should be equivalent, despite different numbers due to how the different code works.