This works
The /taur voidsuits work like a charm.
With the rigsuits, it'll show the correct sprite for taurs while the
rigsuit is sealing, but after it seals up it'll revert to the old
sprite. To fix this, all someone has to do is toggle and retoggle the
chestpiece.

If someone else without a taur/a different taur goes to wear a
rigsuit/voidsuit someone previously wore it'll change to fit their taur.
Pretty neat!

Sprites provided by SilencedMP515
Also, thanks to Nightwing for helping me get this working when I ran
into a brick wall.