Removes Ninja & Malf Voting Malf is a terrible gamemode, the code is unmaintained, and it's not very fun for the crew or the AI.  The AI can still be evil, and more effectively, as a traitor.  At some point I plan to give traitorAIs some malf powers, but the malf mode needs to go, and I meant to do this literally a year ago.
Ninja will no longer be votable, because one ninja cannot carry the round.  The age of solo-antags is coming to an end, because solo-antagonists have been proven to be ineffective, and place massive pressure on the person playing the antag.  It also makes it difficult to balance one person to be able to fight the whole station without it being shit for the station or shit for the antag.  To fix that, Ninja is no longer a solo-antag.  The other ninja modes like Intrigue and Visitors remain, as the ninja should have more breathing room, since other baddies exist.