Fixes dresses
Ports over https://github.com/PolarisSS13/Polaris/commit/301f97a2b388e579fe998a95dce08b5416523834

Instead of syncing (Which would take longer) I'm just manually putting the bugfix in. This will cause a conflict with the next sync, but all that needs to be done is the comment I put in taken out (Along with the heading in github when a conflict occurs) and it'll work fine.

I'll solve the conflict next time we sync next month.