Feature: Extended Character Customization
Added more options to the character customization screen. These options
have dna support, clothes will hide them but can be configured to not,
will only work with races that can use underwear, can be customized
further per player, gender or other variables used by the others
customization systems. I tried to respect as much as I could VR syntaxes
and special areas of code, imitating commentaries, the code suffered
four itterations on the servers I worked, so this one version is already
optimized as much as I could think of. Underwear will also block the
customization, but of course this can be changed with simple vars, and
removing underwear and clothes will reveal it.

Credits for code and port go to OrbisA, credits for sprites go for
BigSky, credit for hairy one goes to Xsdew, credit for robot parts goes
to Tuesday. Additional parts were coded in such a way to be easy to add,
encouraging participation by players.