Makes nanopaste useful externally
Welders and wires repair 15 of brute and burn respectively, this only repairs 5 of both. So you can do it a lot faster with a welder/wires, but you can carry a nanopaste with you to fix that time you hit yourself with a box or something. Quick fieldwork w/o goggles and such.

Fixes #1105