Add vore taste system
There's a setting on bellies (off by default except for the first belly you're ever given which is on by default) as to whether or not they can 'taste' prey. If so, and the prey has set their (literal) flavor text in their vore prefs (button at the bottom) you will be shown their taste when ingested into that belly.

There's also a "Lick Someone" verb in IC that allows you to lick adjacent mobs to see what they taste like.

Both system account for reagents splashed on the mob, so if you want to apply soy sauce to the akula before you eat them, you should be able to taste that as well. Just keep in mind those splashed reagents do dry up eventually, so don't take too long to eat them after that if you actually want to taste it.