Adds 4 waggable tails, *vwag emote, other fixes
Using the *vwag emote you can toggle the vorestation specific tails to wag. Tails marked with (vwag) in the tail selection support this.
New tails are: a new vulp tail (does not replace current one, selectable as a custom tail), a tiger tail, a short tail, and a curl tail (husky or doge). All four support vwag.

Also fixed the mouse tail. It said "mouse, grey" when what it meant was "pink" and then it also said "colorable" mouse tail but that was a lie because it was not colorable at all. Not really. Not in any reasonable way.

Also added fully colorable mouse ears without the pink inner part as an option.