Removes Security
Because apparently we can't have nice things anymore and because so many
people abuse the position of security to play backseat admin, cause
drama, and generally be obnoxious shitters, I've decided that the most
logical solution is to remove security as a number of players have
requested for the past few months. Enough is enough.

Security has been walled off, weapons removed, and the power cut until
as an IC placeholder until we decide what to do with the empty space.