Bluespace Jumpsuit and remove arbitrary size changing
The bluespace jumpsuit is relatively cheap (compared to a shrink ray) and craftable in R&D. When worn, it allows you to resize yourself at will, and when removed you'll snap back to your original size (use a shrinkray or the Mass Altering NIFSoft for permanent changes). The reason it's not permanent is because it's very cheap to make in R&D, really. No mining materials needed, but it only affects you. So you can consider it the 'minor' size change tool.

This PR also comments out the arbitrary size change verb since we decided to remove it, as it was being used to generally just make 'magic' characters that could resize themselves at will, or use it for mechanical purposes to avoid being shot etc when fighting starts, even though it says specifically not to do these things. If you want such a character that can resize themselves due to their physiology, please apply for an SCA and we can give you a small fluff item or other accomidation to allow such a thing.