Adds 3 new guns
* KHI-98a 'Protector'
* MPA6 'Sickshot'
* PML9 'Pummeler'

The Protector is an energy gun that can't be fired on lethal on code green, so it's considered safe for sec to carry regularly, and has a slightly enhanced stun over tasers and normal eguns. It also features a DNA lock that allows one to secure the gun to themselves. An emag can disable both the security level lock and clear DNA.

The Sickshot is a stun weapon in handheld form, that induces nausea and vomiting. It has a chance to make predators vomit out prey (regardless of belly) and tends to cause confusion and nausea. It does cause slight burns from the hypersonic frequency sound it uses, but nothing serious or potentially lethal.

The Pummeler is a 'bass cannon' essentially that fires a huge sound wave, tossing the target away from you. It may or may not knock them down, depending on your luck, but it will definitely push them away, and potentially over a railing or wherever you wanted to put them.