Adds Taur Biosuits to Medical Biosuit Crate
Also this crate is an absolute mess can we split it into like, 2 or 3 crates at a later date? Also aware that the general suit sprite isn't 100% correct, but I figured consistency was better until I finish all of the suits (They don't have different names in game so its not very noticeable).