Rebalanced the Energy Luger
@Yoshax noticed that the Energy Luger can do a retarded amount of damage
over its Energy Gun counterpart. The Energy Luger has 140% more shots on
all settings than the standard Energy Gun. The Energy Luger has the
potential to do 2.4x more damage. This is absurd, and not its intended
design. This was an oversight that occured during a recent rework of
Polaris energy guns and how batteries work.

I have three possible solutions. This is one of them.

A) Rebalance it to match the energy gun, including the ability to reload
it (requires a new sprite which I do not have).
B) Lock the battery, making it impossible to reload without a charge
station, but allow it to keep twice the capacity. (I'll fix the numbers
so it's exactly 2x the Egun capacity, not 2.4x)
C) Lock the battery, but remove the firing delay. **This is what I did
in this PR.**

Discuss the three solutions and decide what is best. Option A is not an
option without a sprite, though, so if anyone wants option A, make a
sprite.