Mobs Lazy Spawner
This is an upgrade of the current mob spawner, what it does is use a
process identical to that of the proximity scanner to check for people
nearby, and if found, it will spawn a mob using the usual rules of the
spawner to get a new mob.
- It wont scan for new users if its on the mob cooldown.
- It comes with a range var to set which tiles should be scanned.

The idea behind this was to prepare a map with the proof of concept, and
launch this lazy loader for the wilderness, since it would fix most of
the lag issues we had given that mobs would only show up when they are
needed.