Item re add: Meta Glasses
Re added the old style metamorphic glasses that changed shape with the
drink it had in them. I left clear examples of how to add new drinks if
needed, as the current itteration only have the original sprites.

Since these glasses are supposed to be rarer, I added the following ways
to get them:

-With one coin, a booze mat machine will be able to vend you ONE.
-Mining abandoned crates have a good chance to get them, and half of
that chance to get a box with four of them.
-Contraband crates will have a decent chance to have a box of 4 of them
-The trashpile has a rare chance (beta) to get a single glass.

This was supposed to be the end quest reward from the Resort Gateway
map, sadly, the project is no longer viable, and thus was dropped, we
still felt the effort shouldnt be wasted, so both this code and the lazy
spawner were released.