Redoes the small cruiser
On the ships map, the little one that was in the top left. Am I the only admin who's ever used the admin lewdship in our maps? This one is too big to fit at the station so it has a little shuttle. Maximum class, for hosting parties.

While editing the docking thing for the shuttle, also fixes the tether 'goodluckmcgee' shuttle docking at the station. Now it docks properly and the doors open like one might expect.