Destroys botany forever
Why do we have a robot that literally does the entire job. Basically people leave it and it becomes a lag generator forever. Removes farmbot harvesting and removal of dead plants, plus adds an idle timer for 150 process cycles without a target causing it to turn itself off. So basically it will tend the plants and keep them alive, but you need to come harvest them sometimes.