Gives the HoS a KHI-102b NSFW It's just my fluff gun. But sharing is caring.
From the ingame fluff:
"Variety is the spice of life! The 'Nanotech Selectable-Fire Weapon' is an unholy hybrid of an ammo-driven energy weapon that allows the user to mix and match their own fire modes. Up to three combinations of energy beams can be configured at once. Ammo not included."
"This gun is an energy weapon that uses interchangable microbatteries in a magazine. Each battery is a different beam type, and up to three can be loaded in the magazine. Each battery usually provides four discharges of that beam type, and multiple from the same type may be loaded to increase the number of shots for that type."
"The Kitsuhana 'Nanotech Selectable Fire Weapon' allows one to customize their loadout in the field, or before deploying, to achieve various results in a weapon they are already familiar with wielding."

The HoS's includes 2 lethal, 2 stun, 1 net, and 1 ion battery. There are other batteries (like instagib!) but there's no way to get them normally. There's also one that strips people.