Tweak lethal succubus drain
Mostly remove the OOC magic-implying wording about 'life energy' (since setting is w/o magic) so it can be whatever in your private RPs (someone was asking about using it to represent stealing bodyheat which is cool).
`absorbing_prey` was never unset if they were found to be indigestible, so fixed that
Two stages also sent two messages at the same time
Removed the needless casting of src to C since it's defined on human already

Maybe in the future consider using cloneloss instead of brainloss? Alkysine is one of the easiest drugs to make, and if I jack myself up with it I'd maybe be immune to having my nutrition/heat/whatever stolen which seems ~unusual~