Redoes the armory
Space Obama here, to take all the guns away. This removes the entire secure side of the armory and replaces it with a shuttle with guns and heavy armor that NT can send if the station is at code red. It's in the same place, but has fewer guns oh noooo. The console to call it is in the not-as-gunny side of the armory, but still has access to both sides (from the breakroom and from the other side of the armory).