Prevent stalls in dogborg self-clean cycle
When something unexpected (such as mice, which are neither humans nor
items) is in a scrubpup's garbage compactor, it could stall forever when
performing a self-clean. This changes it to treat anything it doesn't
know how to digest as an indigestible item, treating it like other
indigestible items such as ID cards.

This also changes the self-clean damage ticks to harm anything living,
rather than just human.