POLARIS: Simple Animal Player QOL
Adds a HUD when controlling a simple animal. Allows intents, mostly harm and help if they don't have hands. Disarm and grab if they do. Gives them a generic hud, and adds a var to decide if they have hands or not. Adds a var printed to any client that joins into a mob to explain how to play them, or lore stuff, or whatever you feel like.