Readds NIF examine text
Recoded it a little different. The verb is in OOC, if you have a NIF. It's saved as part of the NIF, so if your head explodes and you lose your NIF, you'll need to describe your new one.

Also fixes an unrelated bug where ghosts examining people print all sorts of crazy messages about being able to trample people.