44 revolvers
Port from Bay

- Addition of two more revolvers in 44 cal for variety
- Creation of rubber rounds for 44 cal (stats taken from virgo no idea how good they are in actual fights)
- Includes 44 ammo speedloaders, regular and rubber speedloaders
- Slight adjustments to allow for firing animations taken from Bay (the mosley revolver has a firing animation that plays)
- Includes in hand sprites

We did already have the 'webley' sprite (it exists in a recolored sprite). However it is a alternate sprite for the detective's 45 revolver. This brings it in as a actual proper revolver with sprites that are different than the detectives revolver, and maybe in the future can be thrown into loot lists and PoIs.