Makes donuts less of a strain on resources.
considering you can put 6 donuts in a box, and they're supposed to be a small food item... this overrides all the donuts to use dough slices instead of a full wad of dough. 

Literally, who would realistically use two boxes of flour, six eggs, and whatnot to make one batch of donuts?