Adds the Resonator
A device which sets up resonating energy fields which collapse, mining mineral turfs and/or doing damage to any mobs standing inside.  Does no damage to anything else, and damage to mobs is only high when used in low temperature environment (such as... outside or in space)
Uses quite a few dmis, for the icons, in hand icon, and effect icons.  Includes sound effects
Ported from tg and Paradise