Tails/Taurs Hide Shoes & Clip Uniforms/Suits
Hides shoes when the goat legs or taur bodies are selected.
Adds a method by which you can clip off parts of sprites, ported from Para.

Sets goat legs & taur bodies to clip off the lower parts of uniforms and suits
so long as the clothing doesn't hide tails or is a taur suit.