Chapel Nitpicks
Re-adds medical access to the morgue's back door, adds a tintable window and button to noodle's office, moves the disposal bin back so that you can actually get around the desk. Puts the little bit of maint loot back. And trades a locker for the cyborg charger to make the chapel access airlock 2 wide again.