Centcom/Tether Ship Changes
- Centcom arrivals changed to south, instead of west.
- Removed outdoor areas from Centcom
- Moved Aronai and Joan's houseboat to admin_use

The goal here was to open up some space at the CentCom Z level for future things down the line, which is why the CC Z level now has a 140x50 chunk of empty space.

In the future (either in this PR, or a future PR) I will be readding an admin ship to the Centcom Z level in that empty space that will functionally serve as the replacement for the old houseboat, as well as using that Z level for other projects.