Fixes chairs acting as handcuffs. (#6588)
* Fixes chairs/beds/etc acting as handcuffs.

It's been at least 7 years without anyone noticing a broken feature with a false comment so the unintentional "balancing" caused by this may have ingrained a standard some folks would rather keep in, but that kind of nerfs have proper ways of implementing at their own appropriate places in the code.

Basically this just lets folks use hands when buckled. May also have a side effect of allowing self-surgery if it never had a proper prevention method implemented. Keeps an exception for xeno nests tho. Any other seat has no reason to be handcuffs.

* Fixes another thing causing chairs to incapacitate

"Are you seated" and "Are you incapacitated" should not be used as the same thing yet here we are.

I should probably also do a refactor to remove all the inappropriate use of canmove var. Seriously, even the specifically cartridge exclusive PDA menu buttons are using that for what???