Huge Map Tweaks
S1:

Made the pillars outside R-wall. They support a portion of the station, it only makes sense that they would be ultra sturdy.

Fixed a... HUGE portion of Surface Science to have R-wall. Good god, any greytider with a welder could have busted in in seconds before this change.

Added R-wall along atmos's eastern S1 wall to prevent busting in through holodeck dorms.

Added a box of Replacement lights to Surface med, overhauled the resleeve shower room to be public access, actually have a shower and not just a curtain with drains, and added some towels and bathrobes.

Fixed some issues with the Janicloset, making both doors say the same thing with the same color code, fixing the interier windoor under the shutters to be jani access only [previously public] and added a third bottle of space cleaner [one for each janitorial cart]

Added a light in the Maint tunnels so that the emergency evacuation from Surface Med to the Panic Shelter is fully [albeit dimly] lit

Continued to secure Science with R-walls, though not nearly as much as was needed on S1.

S3.

Secured the outer walls of Xenobio and the isolation hallway with R-wall to match the R-windows, put the finishing touches on R-walling Science and Robotics

Patched a few r-wall holes in Surface sec [nothing major, just some corners needed reinforced to match the rest of the walls]

Adjusted the Tether automated backup shuttle to and from the station to be operable by the public. It's an emergency shuttle for goodness sake, having only exploration and Pilot's being able to operate it was a major safety hazard in the event the elevator was non-operational! [Modified all three control consoles to do this, S3, Shuttle, and A1]

Fixed lighting issues in the Panic Shelter.

Changed steel tile to wood tile in public study [look under the doooooors]

A1

Added R-walling along backup atmospherics to lessen the damage of canister explosions to Engineering's EVA and tool storage areas.

Added a way for engineering to lift Primary engineering lockdown from the outside [like they can do with atmos]

Added an Engine Room lockdown separate from Engineering lockdown to combat greifing. Greytide will now be required to gain access to Engine Room level clearance to lift blast doors.

A2

Added a way for engineering to lift Primary engineering lockdown from the outside [like they can do with atmos] 2, Electric boogaloo

Added R-walling to Tech storage and secure tech storage.

A3

Added R-walling along the outside of Sec that was missing in some areas, as well as fully R-walling Forensics.

Mining:

Added a trashbin to combat litterbugs. Offset fire closet to adjust for bin placement.

Rewired exterior APC to go under the window instead of the wall