Moving stuff for fluff
Moved some lights and items around in the surface medbay to make it less clutterd. Also added a sink.

Spun some wheelchairs around in Asteroid medbay storage so the backs are facing the wall.

Should be my final commit for this PR until it's either accepted or I manage to complete moving the CMO's office to Surface.