Buffs ERT Backpack
I always thought it strange that the ERT had both their own backpacks, and access to the unquestionably superior Syndicate dufflebags.

This buffs the ERT Backpack to be on par with the syndicate dufflebag. Long term plan is to remove the Syndicate dufflebags from the ERT Armory, but that cannot be done yet due to the mapping freeze.