Adds Vespa
A ship meant for use within events.
This is a vessel for exploration purposes.
It is meant to be an escape for event-goers. This is a place where events may happen at any time, without warning. People that step foot on this ship must recognise that their RP/ERP will be interrupted by any number of things. Of course, during quiet times, they can do w/e.
First part of map - another PR will come with an updated Vespa dmm AFTER an event.