Attempted mapmerge + additions
Adds the following
The vespaboat! A functioning shuttle that allows players to move between landmarks. Is NOT an overmap shuttle.
Adds functioning airlocks. North and south + the shuttle's airlock.
Adds the supermatter shard to engineering. Utilizes radiation collectors. The room is made out of lead, so radiation shouldn't be too much of a problem!
Removed security. All of it. Will place armoury in another spot later on. Maybe.