Reenable Air Pump Soundloop
After extensive testing, there's no way to simply confine the airpump to a single sound channel without extreme wonkiness/inaudible audio, and the suggestion of a global audio channel resulted in no audio playing at all.

Attempting to make a static loop caused horrible feedback, so that was a no-go, and attempting to set up a check to find out if there are mobs near a vent (basically a proximity check) does the same thing/uses the same amount of resources that simply doing playsound() does - but in reverse.

Tested with ShadowLarkens <shadowlarkens@gmail.com>'s help