Fix language keys not working after multilingualism
Sort of a "fix", sort of a tweak; the system deliberately never checked
for anything but the lowercase version of a language key to make it
easier to use, but a lot of the VOREStation languages use uppercase
invocation keys. Easy enough to fix by just making it check for the
literal version of the key instead of the lowertext version.