Beginning of adding a Cyborg job slot to Talon
Adds a Cyborg job slot to Talon crew, same 14 day limit as the rest of the crew, and adds a spawn/leave storage spot as well as a recharger next to it.

Modules are Tether default, access is Tether default, I'll have to figure out how to change that, because a lot of the Cyborg/AI stuff is hardcoded to assume there's only ever one main "station", it seems like?

I'm very new to this, so I'm probably doing it wrong.