Adds Equine Facial Markings & Ears
I was poring over the markings list as usual, inspired this time to draft a donkey but was woefully struck by a lack of equine markings. Neigh, this cannot be! So I took to resolving the issue post-haste.

The muzzle is flexible goes quite well with draconic designs too. Recommended to use with the Sergal bodytype for best results.