New overmap POI
Adds a new map called "Away Mission - Guttersite"
An installation built on a group of asteroids in VE space serves the purpose of giving gutter speaking folk a place to do their shady dealings outside of solcomm jursidiction.
This map is coded up to show up on the overmap, and is accessible by both the excursion shuttle and talon boat.
The map is full of secrets and content, and leaves a lot of room for staff to fill it up with unique mobs and people for events.