Machines can't be used by clowns.
The old system was unfair, only the carbons or the silicons amongst us were allowed to use computers or see the internet. Or adjust the thermostat. But no longer. But changing this area here, it lets any mob with "has_hands = TRUE" use the objects they should be allowed to..