NanoUI tweak (#8268)
* Machines can't be used by clowns.

The old system was unfair, only the carbons or the silicons amongst us were allowed to use computers or see the internet. Or adjust the thermostat. But no longer. But changing this area here, it lets any mob with "has_hands = TRUE" use the objects they should be allowed to..

* simple_mob Revolution

Allows simple mobs to interact with nanoUI if they have "have_hands = TRUE".

* Revert "Machines can't be used by clowns."

This reverts commit 3cb16f57c0986e492c8f247598fc74773cebea63.

* Moves patch.

Moves Patch to default_vr.dm

* Adds missing part to machinery fix.

Fixes my dummy mistake and effectively unreverts the original commit.