Another service tweak.
Changes the name of the entertainment room airlocks to "Auditorium".
Changes up seating in the auditorium.
Cuts down on the size of the auditorium stage so anyone seated can see anything on the stage. There is a strip of area on the ends for placing the spotlight as well.
Addsa megaphone to the entertainment backroom.
Adds intercomms across the service block.