Fixes Nanotrasen Uniforms to account for Solgov Path Changes
This was introduced thanks to a PR in mid-2019, and NT Sec/etc uniforms had been broken since then.

Screenshot of broken uniforms:
![](https://i.imgur.com/AVii80Z.png)

As of now, the in-hand sprite is the only one that's 'broken', display JSUIT on your sprite, but the actual UI in-hand works, the on-ground works, the on-mob works, so frankly, after two+ hours, I'm calling it good here.

This fixes all derivative NT uniforms too.