Fixes conflicts with incoming Mirror #8756, Re-enables ecigs
At this point, ecigs have been arbitrarily disabled for who knows how long for... no idea why, honestly?

I've added the missing .dmi and re-added ecigs to the .dme, in the interest of not conflict-fixing a million problems with vending machines.

E-cigs will need to be added to loadouts manually later.