[WIP] Torris Jungle Base (Name TBD for Base!)
Welcome to the [PENDING NAME] Base on Torris!

An extremely warm, humid jungle planet in [random system] with a large, nanotrasen base in one of the larger jungles!

Your base is above the treetops, with a ground 'vehicle' area and small complex at the very base. Hungry panthers and other jungle critters wander around, so venture beyond the fence at your own peril!

View current doodles here:
Verkister V1:
![](https://cdn.discordapp.com/attachments/187013248309002240/746700847500492870/4c0eb32368761f105fe709bfbcc86d7f.png)
Verkister V2:
![](https://cdn.discordapp.com/attachments/187013248309002240/747119854615527544/6f80b31bfb90139e8deb077f95248189.png)
Scree:
![](https://cdn.discordapp.com/attachments/315701453463486464/733049358387511386/unknown.png)

This is a DRAFT and is intended more as a show of progress and to allow collaborators to start working on the project!